+++ crates/miden-testing/src/mock_chain/chain_builder.rs
@@ -157,7 +157,7 @@ impl MockChainBuilder {
         let nullifier_root = NullifierTree::new().root();
         let note_root = note_tree.root();
         let tx_commitment = transactions.commitment();
-        let tx_kernel_commitment = TransactionKernel::kernel_commitment();
+        let tx_kernel_commitment = TransactionKernel.to_commitment();
         let proof_commitment = Word::empty();
         let timestamp = MockChain::TIMESTAMP_START_SECS;
         let fee_parameters = FeeParameters::new(self.native_asset_id, self.verification_base_fee)